+++ Kingdom4/src/data/IEvent.java
@@ -1,11 +1,11 @@
-package game;
+package data;
 
 import javafx.geometry.Point2D;
 
 /**
  *
  * This defines the structure of an event field, e.g. collision or trigger.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
