+++ Kingdom4/src/game/CollisionDetection.java
@@ -5,11 +5,11 @@ import javafx.scene.shape.Rectangle;
 /**
  *
  * This detects collisions in the game world.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
- * @version 1.0
+ * @version 1.2
  *
  */
