+++ Kingdom4/src/game/Hero.java
@@ -9,11 +9,11 @@ import javafx.scene.shape.Rectangle;
 /**
  *
  * This is to worship the hero we at Fantastic 4 studios love and appreciate.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
- * @version 1.0
+ * @version 1.2
  *
  */
