+++ Kingdom4/src/game/IObservable.java
@@ -3,7 +3,7 @@ package game;
 /**
  *
  * This defines the structure of the Observable class that implements the methods below.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
