+++ Kingdom4/src/game/Kingdom4.java
@@ -1,5 +1,8 @@
 package game;
 
+import data.World;
+import graphics.Hud;
+import graphics.Scenery;
 import javafx.animation.AnimationTimer;
 import javafx.application.Application;
 import javafx.stage.Stage;
@@ -7,7 +10,7 @@ import javafx.stage.Stage;
 /**
  *
  * This makes the magic happen.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
