+++ Kingdom4/src/game/Observable.java
@@ -6,7 +6,7 @@ import java.util.List;
 /**
  *
  * Implements the interface IObservable and provides methods to register and notify observers.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
