+++ Kingdom4/src/game/TriggerDetection.java
@@ -1,15 +1,16 @@
 package game;
 
+import data.Trigger;
 import javafx.scene.shape.Rectangle;
 
 /**
  *
  * This detects triggers in the game world.
- * Copyright (c) 2018 Fantastic 4 Studios. All Rights Reserved.
+ * Copyright (c) 2018-2019 Fantastic 4 Studios. All Rights Reserved.
  * @author Fabian Schmidt
  * @author Martin Sanfilippo
  * @author Boris Bischoff
- * @version 0.9
+ * @version 1.0
  *
  */
